+++ RELEASE_NOTES.md
@@ -20,6 +20,8 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
   - A stored device ID that is currently unavailable will now appear as an "Offline Device" instead of "No Device," differentiating between having no device selected and an unavailable saved device.  
 - **ZWO EAF Native Driver**  
   - The ZWO EAF is now available as a native driver.  
+- **New SVBony Camera Driver**
+  - Added native support for the new SVBony camera series 
 - **ASCOM Device Name Query**  
   - N.I.N.A. now queries the ASCOM device name after connecting to the driver.